requirements.txt
Django==1.9.5
Jinja2==2.8
requests==2.9.1
mandrill==1.0.57
django_extensions==1.6.7
mailchimp==2.0.9
pytz==2013.7
psycopg2==2.6.1
djangorestframework==3.3.3
drfdocs==0.0.9
django-crispy-forms==1.6.0
django-push-notifications==1.4.1
passbook==1.0.2
django-sql-explorer==1.0
qrcode==5.3
django-boto==0.3.9
